fix: fix funtion name for PR #34
