--- src/utils/isWithdraw.ts
@@ -1,20 +1,0 @@
-import { log } from 'console';
-import dateToSeconds from './dateToSeconds';
-
-const isWithdraw = (
-  startDate: number,
-  cliffDate: number,
-  amount: number,
-  withdrawn: number,
-  isCancelled: boolean,
-) => {
-  const currentDate = dateToSeconds(new Date());
-
-  if (cliffDate > currentDate && currentDate < startDate && withdrawn === amount && isCancelled) {
-    return true;
-  }
-
-  return false;
-};
-
-export default isWithdraw;
